LPS-74839 petra-json-web-service-client - TEST - Reorganize - change package to server.simulator
