+++ modules/apps/foundation/petra/petra-json-web-service-client/src/test/java/com/liferay/petra/json/web/service/client/server/simulator/HTTPServerSimulator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.petra.json.web.service.client.serversimulator;
+package com.liferay.petra.json.web.service.client.server.simulator;
 
 import com.sun.net.httpserver.HttpHandler;
 import com.sun.net.httpserver.HttpServer;
